fix(tests): Move displayAndUpdateWorkItemDetails to Excluded folder
